# compiler: avoid allocation in SetJumpTargetOnNext