_arm64_

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                    │   old.txt   │              new.txt              │
                                    │   sec/op    │   sec/op     vs base              │
Compilation/with_extern_cache-10      132.6µ ± 1%   133.0µ ± 0%       ~ (p=0.165 n=7)
Compilation/without_extern_cache-10   2.225m ± 0%   2.188m ± 1%  -1.68% (p=0.001 n=7)
geomean                               543.2µ        539.4µ       -0.71%

                                    │   old.txt    │              new.txt               │
                                    │     B/op     │     B/op      vs base              │
Compilation/with_extern_cache-10      53.40Ki ± 0%   53.40Ki ± 0%       ~ (p=0.511 n=7)
Compilation/without_extern_cache-10   1.159Mi ± 0%   1.144Mi ± 0%  -1.28% (p=0.001 n=7)
geomean                               251.8Ki        250.1Ki       -0.64%

                                    │   old.txt    │               new.txt               │
                                    │  allocs/op   │  allocs/op   vs base                │
Compilation/with_extern_cache-10        979.0 ± 0%    979.0 ± 0%       ~ (p=1.000 n=7) ¹
Compilation/without_extern_cache-10   10.384k ± 0%   9.414k ± 0%  -9.34% (p=0.001 n=7)
geomean                                3.188k        3.036k       -4.79%

```

_amd64_ (with_extern_cache must be noisy)

```
benchstat old_amd64.txt new_amd64.txt
goos: darwin
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: VirtualApple @ 2.50GHz
                                    │ old_amd64.txt │           new_amd64.txt           │
                                    │    sec/op     │   sec/op     vs base              │
Compilation/with_extern_cache-10        327.0µ ± 2%   333.6µ ± 2%  +2.02% (p=0.004 n=7)
Compilation/without_extern_cache-10     4.792m ± 1%   4.781m ± 1%       ~ (p=0.456 n=7)
geomean                                 1.252m        1.263m       +0.90%

                                    │ old_amd64.txt │           new_amd64.txt            │
                                    │     B/op      │     B/op      vs base              │
Compilation/with_extern_cache-10       53.64Ki ± 0%   53.64Ki ± 0%       ~ (p=0.557 n=7)
Compilation/without_extern_cache-10    1.168Mi ± 0%   1.153Mi ± 0%  -1.29% (p=0.001 n=7)
geomean                                253.3Ki        251.6Ki       -0.65%

                                    │ old_amd64.txt │           new_amd64.txt           │
                                    │   allocs/op   │  allocs/op   vs base              │
Compilation/with_extern_cache-10         981.0 ± 0%    981.0 ± 0%       ~ (p=0.462 n=7)
Compilation/without_extern_cache-10     11.33k ± 0%   10.36k ± 0%  -8.55% (p=0.001 n=7)
geomean                                 3.334k        3.189k       -4.37%

```